+++ include/hyperliquid/hyperliquid.hpp
@@ -4,6 +4,7 @@
 #include <hyperliquid/utils/constants.hpp>
 #include <hyperliquid/utils/types.hpp>
 #include <hyperliquid/utils/keccak.hpp>
+#include <hyperliquid/utils/l2_diff.hpp>
 #include <hyperliquid/utils/signing.hpp>
 #include <hyperliquid/api.hpp>
 #include <hyperliquid/websocket_manager.hpp>